Move GDS cuFile context into gpu_connector package
