+++ lmcache/v1/gpu_connector/gpu_ops.py
@@ -3,9 +3,9 @@
 import torch
 
 # First Party
+from lmcache.v1.gpu_connector.gds_context import get_gds_context
 from lmcache.v1.lazy_memory_allocator import LazyMemoryAllocator
 from lmcache.v1.memory_management import GDSMemoryObject, MemoryObj
-from lmcache.v1.multiprocess.gds_context import get_gds_context
 import lmcache.c_ops as lmc_ops
